extracted TextInput component to reuse inside ListSearchField
